+++ CHANGES.md
@@ -11,6 +11,9 @@
 
 ### Fixed and improved
 
+* Update Java to 8u192. (DCOS_OSS-4380)
+
+* Docker-GC will now log to journald. (COPS-4044)
 
 ### Notable changes